Create .gitignore
These are filetypes that should NEVER be in the tree, and, if by accident they are there, not included in the commission.  Also, I use, on occasion, tools to generate, format, or otherwise polish my scripts, usually from C++ code i made myself, but since there should be no binaries included (generally speaking), then these files must be ignored, even if I keep them inside the directory locally. I suggest anyone else that choses to work with the scripts to do the same. Also, vim likes to make .swp files which is good, but not for the tree, so they must be ignored since they have no meaning on another machine. Furthermore, no files ending in ~ will be permitted as it is a general name for UNIX backup files that is usually skipped by most file browsers anyway.