validated README for forkers
updated the readme documentation because it contained incorrect info about installing. the install instructions have been finalized to fit the depreciated subject matter. I have tested these myself on a separate system and user to verify it works. Be warned that gxbase is not finished so dont expect it to work. The tools for the most part are usable but the subshell is not. See autox for the current project using this projects code and tools!